+++ index.html
@@ -64,6 +64,6 @@
       integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4"
       crossorigin="anonymous"
     ></script>
-    <script src="js/main.js" type="module"></script>
+    <script src="js/index.js" type="module"></script>
   </body>
 </html>
