+++ js/index.js
@@ -1,0 +1,5 @@
+import { buildMenu } from "./ui/common/buildMenu.js";
+import router from "./router.js";
+
+router();
+buildMenu();
