+++ js/listeners/editPost.js
@@ -1,8 +1,8 @@
 import { BASE_URL } from "../api/constants.js";
-import { accessToken } from "./token.js";
-import { getPostData } from "../handlers/getPostData.js";
+import { accessToken } from "../api/token.js";
+import { getPostData } from "../api/posts/getPostData.js";
 import { getPostId } from "../handlers/getPostId.js";
-import { updatePost } from "../handlers/updatePost.js";
+import { updatePost } from "../api/posts/updatePost.js";
 
 const form = document.querySelector("#editPostForm");
 const id = getPostId();
