+++ js/listeners/index.js
@@ -1,0 +1,4 @@
+export * from "./login.js";
+// export * from "./editPost.js";
+// export * from "./register.js";
+// export * from "./showPostByTags.js";
