+++ js/listeners/register.js
@@ -1,5 +1,5 @@
 import { BASE_URL } from "../api/constants.js";
-import { fetchData } from "../api/call.js";
+import { fetchData } from "../api/fetchData.js";
 
 const form = document.querySelector("#registerForm");
 const endpoint = "auth/register";
