--- js/main.js
@@ -1,3 +1,0 @@
-import { buildMenu } from "./ui/buildMenu.js";
-
-buildMenu();
