+++ profile/login.html
@@ -23,7 +23,7 @@
           <span class="navbar-toggler-icon"></span>
         </button>
         <div class="collapse navbar-collapse" id="navbarSupportedContent">
-          <ul class="navbar-nav me-auto mb-2 mb-lg-0">
+          <ul class="navbar-nav me-auto mb-2 mb-lg-0" id="menu">
             <li class="nav-item">
               <a class="nav-link" aria-current="page" href="/">Home</a>
             </li>
@@ -104,6 +104,6 @@
       integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4"
       crossorigin="anonymous"
     ></script>
-    <script src="../../js/handlers/login.js" type="module"></script>
+    <script src="../js/index.js" type="module"></script>
   </body>
 </html>
